Added Project Refinement

+++ src/main/java/com/retail/reward/model/Transaction.java
@@ -1,14 +1,10 @@
 package com.retail.reward.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
- * Represents a customer's purchase transaction.
- * * @param id The unique identifier for the transaction.
- *
- * @param customerId      The unique identifier of the customer making the purchase.
- * @param amount          The total transaction amount in dollars.
- * @param transactionDate The date the transaction occurred.
+ * Represents a customer's purchase transaction in the domain layer.
  */
-public record Transaction(Long id, Long customerId, Double amount, LocalDate transactionDate) {
+public record Transaction(Long id, Long customerId, BigDecimal amount, LocalDate transactionDate) {
 }
